parsing comp

+++ libft/ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: jeykim <jeykim@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/07 15:42:28 by jeykim            #+#    #+#             */
-/*   Updated: 2022/04/02 13:10:48 by jeykim           ###   ########.fr       */
+/*   Updated: 2022/04/09 21:47:07 by jeykim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
